Organize scripts into folders.
