+++ resources/assets/js/utilities/confirm.js
