+++ resources/assets/js/utilities/method-link.js
